requirements.txt
fastapi
uvicorn
python-multipart
boto3
python-dotenv
httpx
beautifulsoup4
requests
PyPDF2
python-docx
chardet
markdown 